feat: Add CORS configuration, implement findByEmail, and enhance logging
- Added CORS configuration to allow requests from specific origins.
- Implemented findByEmail function in PrismaUserRepository to retrieve user details by email.
- Moved user-related endpoint to a more appropriate location for better organization.
- Added detailed loggers to improve traceability and debugging.
